base dao service

+++ cloud-note/src/main/java/com/sxkl/project/cloudnote/article/service/ArticeService.java
@@ -22,5 +22,5 @@ public interface ArticeService {
 
     Page<Article> findPage(int page, int size);
 
-    int countByNoteId(String noteId);
+    long countByNoteId(String noteId);
 }
